+++ Core/Math/Quat.cs
@@ -3,7 +3,7 @@ using System;
 namespace FIVES
 {
     public struct Quat {
-        public float x, y, z, w;
+        public double x, y, z, w;
     }
 }
