+++ Plugins/Location/LocationPluginInitializer.cs
@@ -53,9 +53,9 @@ namespace LocationPlugin
         {
             // Position is represented as a vector (x,y,z) from the default position, which is at (0,0,0).
             ComponentDefinition position = new ComponentDefinition("position");
-            position.AddAttribute<float> ("x", 0f);
-            position.AddAttribute<float> ("y", 0f);
-            position.AddAttribute<float> ("z", 0f);
+            position.AddAttribute<double> ("x", 0.0);
+            position.AddAttribute<double> ("y", 0.0);
+            position.AddAttribute<double> ("z", 0.0);
             ComponentRegistry.Instance.Register(position);
 
             // Orientation is represented as a quaternion, where (x,y,z) is a vector part, and w is a scalar part. The 
@@ -63,10 +63,10 @@ namespace LocationPlugin
             // orientation, the viewer is on the Z-axis looking down the -Z-axis toward the origin with +X to the right 
             // and +Y straight up.
             ComponentDefinition orientation = new ComponentDefinition("orientation");
-            orientation.AddAttribute<float>("x", 0f);
-            orientation.AddAttribute<float>("y", 0f);
-            orientation.AddAttribute<float>("z", 0f);
-            orientation.AddAttribute<float>("w", 1f);
+            orientation.AddAttribute<double>("x", 0.0);
+            orientation.AddAttribute<double>("y", 0.0);
+            orientation.AddAttribute<double>("z", 0.0);
+            orientation.AddAttribute<double>("w", 1.0);
             ComponentRegistry.Instance.Register(orientation);
         }
